backend/requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector,elevenlabs,groq]~=1.2
livekit-plugins-noise-cancellation~=0.2
python-dotenv
requests
socketio
pymongo
motor
